Correcting physical delivery of Sav. Eats
Definition of physical delivery is a store needs the user to be present at a certain place to receive his product. Check the definition by hovering the mouse over the field... Might be a good idea to change the name to "Digital goods" and change the "yes" to "no" and vice-versa, that way its less confusing.